Seccion 7 => Clase: Pruebas-SaveFile UseCase custom values

+++ src/domain/use-cases/save-file.use-case.test.ts
@@ -2,12 +2,26 @@ import { SaveFile } from './save-file.use-case';
 import fs from 'fs';
 describe('SaveFileUseCase', () => { 
 
-   /*  beforeEach(()=>{
+    /* beforeEach(()=>{
         fs.rmSync('outputs', {recursive: true});
     }); */
 
+    const customOptions = {
+        fileContent: 'custom content',
+        fileDestination: 'custom-outputs/file-destination',
+        fileName: 'custom-table-name',
+    };
+
+    const customFilePath = `${customOptions.fileDestination}/${customOptions.fileName}.txt`;
+
     afterEach(()=>{
-        fs.rmSync('outputs', {recursive: true});
+
+        const outputFolderExists = fs.existsSync('outputs')
+        if (outputFolderExists) fs.rmSync('outputs', {recursive: true});
+
+        const customOutputFolderExists = fs.existsSync(customOptions.fileDestination);
+        if(customOutputFolderExists) fs.rmSync(customOptions.fileDestination, {recursive: true});
+        
     });
     
     test('should save file with default values', () => { 
@@ -27,8 +41,21 @@ describe('SaveFileUseCase', () => {
         expect( result ).toBeTruthy();
         expect(fileExists).toBe(true);
         expect(fileContent).toBe('test content');
-
         
      })
-    
- })
+
+     test('should save file with custom values', () => {
+
+        const saveFile = new SaveFile();                
+
+        const result = saveFile.execute(customOptions);
+        const fileExists = fs.existsSync(customFilePath);
+        const fileContent = fs.readFileSync(customFilePath, {encoding: 'utf-8'});
+
+        expect(result).toBe(true);
+        expect(fileExists).toBe(true);
+        expect(fileContent).toBe(customOptions.fileContent);
+
+    });
+
+ });
